feat(adieu-razzle): moved paths.js into build-utils directory
