+++ packages/falcon-client/build-utils/webpack/config/env.js
@@ -1,8 +1,8 @@
-const paths = require('./paths');
 const fs = require('fs');
+const paths = require('./../../paths');
 
 // Make sure that including paths.js after env.js will read .env variables.
-delete require.cache[require.resolve('./paths')];
+delete require.cache[require.resolve('./../../paths')];
 
 const { NODE_ENV } = process.env;
 if (!NODE_ENV) {
